requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.18
django-celery-results==2.3.1
django-crontab==0.7.1
django-debug-toolbar==3.4.0
django-dotenv==1.4.2
django-environ==0.8.1
django-filter==22.1
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
hiredis==2.2.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
msgpack==1.0.4
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
prompt-toolkit==3.0.36
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dotenv==0.20.0
pytz==2022.7.1
PyYAML==5.3.1
redis==4.3.4
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
wrapt==1.14.1
zipp==3.13.0
zope.interface==5.5.2
django-taggit==1.2.0